fix token registration
