+++ bwell-kotlin-android/app/src/androidTest/java/com/bwell/sampleapp/MyActivityTest.kt
@@ -5,7 +5,6 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.matcher.ViewMatchers
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.bwell.sampleapp.activities.NavigationActivity
